Implemented another approach to handle hardware breakpoints, etc.
The implementation is experimental.

1. HW breakpoints can now be set and cleared from any context (process,
IRQ, softirq, ...). The basic
API is now provided (see hw_bp_set() and hw_bp_clear).

On the current CPU the HW BP is set and cleared directly, timers are
scheduled on the remaining ones to do so there.

Part of the implementation is based on the code from Perf and KGDB.

2. System-wide workqueues are now used instead of the custom one.

3. Got rid of the addr_timer, replaced it with a delayed work.

4. Many minor cleanups and fixes have also been made.