'breakpoints' file now shows what was written to it
Before this commit, when the 'breakpoints' was read, it showed the list
of the software breakpoints actually set at the moment (0xcc present at
the relevant position in the target's code).

The user could write to that file to add/remove breakpoint available for
use (or a range of them).

However, there was no way to see all the breakpoints the user specified
this way. Moreover, if a breakpoint was in the process of being reset
or was not chosen for setting, it would not be shown when reading the
file.

This may lead to confusion: the user might think that the system failed
to add the breakpoint (I thought so myself ;-) ).

I suppose, it is better to make 'breakpoints' file consistent w.r.t.
reads and writes and make reads show the elements of 'ranges_list',
that is, exactly what was specified by the user so far.

Besides that, a release() operation was added to the file to release
the memory allocated by open().

The file was made non-seekable as well, just in case.

Some minor changes were also made.