Handled the case when the code wakes up on a different CPU
After msleep(), the code may wake up on a different CPU. It is unclear
if the pending timer to set the HW BP on that CPU has run its function
by then, however unlikely it seems.

Now hw_bp_clear() takes care to delete all pending timers that set the
BP, on each CPU, including the current one. So even if setting of the
HW BP was pending on the current CPU, it will be cancelled properly.