(uint)(-1) is now used internally to denote "all possible offsets"
If the user writes "func+*" to 'breakpoints' file to specify all
possible instructions in "func", -1 is now used internally as the value
of the offset to denote this (rather than 0).

It is rare but possible for the first instruction in a function to be
of interest to RaceHound too, so using 0 as a special value of the
offset is undesirable.