Warn about a missing function only if target is loaded
One of the convenient abilities of RaceHound is that the user may
specify the desired instructions in the target to monitor before the
target is loaded. As soon as it is, they will be processed.

In this case, I suppose, RaceHound should not warn about a missing
function when the target is not loaded, to avoid confusion.