Revisited handling of software BPs and of the related data
1. The instances of sw_used struct can no longer disappear at a whim.
They are refcounted now. Even if the user requests to delete them or the
code that chooses the BPs to be active removes them, it should no longer
be a problem to their handlers.

Without this, I ran into use-after-free conditions on such structures
for this reason before (during the analysis of e1000 driver).

2. No need to copy function name, etc. all the time. sw_active has been
removed too.

Now the name of the function is stored in sw_available and sw_used just
refers to it.

The fields of sw_active were moved to sw_used. These structures can be
put to two lists, used_list and active_list, depending on the situation.

One entity less.

3. Fixed some memory leaks: the list items were not always destroyed.