Decreased the default delay to 5 jiffies (expressed in msecs)
The "detour buffer" mechanism currently used allows to avoid resetting
the software BPs often. This way, smaller delays should be enough to
detect races.

If 'delay' parameter is 0, the default value of jiffies_to_msecs(5) will
be used. In the experiments with the self-tests for RaceHound this
proved to be more than enough: twice as more races were then found in
the same or less amount of time than before.