The insn decoder can now be used by both kernel- and user-space parts
The decoder may be needed for the tool to filter insns to be watched
by the kernel part later. This should be handy in the scenario where
the instructions to be watched in the kernel are somehow determined
in user space in advance.

The scheme, as I see it now, is as follows.

1. A GCC plugin or some other tool produces the list of source locations
for the memory accesses to be tracked later (file:line or so) when the
kernel and/or the modules of interest are built.

2. Another tool (lines2insns?) uses the binary code of the kernel and
the modules, along with the debug info to find the addresses of the
insns to be watched in the form "[module:]{init|core}+0xoffset".

During this process, it additionally filters out the insns that do not
access memory but happen to correspond to the same source lines as the
memory accessing insns. It may also filter out %rsp/%esp-based accesses,
locked ops, etc.

For all that, the instruction decoder is quite convenient.

3. The insns from the list created at step 2 are fed to the kernel-mode
part of RaceHound. Note that another user-space application may be
prepared to manage when to set software breakpoints, for how long, etc.