Removed section-related stuff, user-mode helper, etc.
All that was not used anyway.

Besides, it is much more difficult now to obtain the addresses of the
module's sections because they are available in sysfs only after the
module has been loaded and notifier functions completed.

Better to stick to init+0xoffset and core+0xoffset in the future:
it is more reliable.

While at it, I removed a couple of other unused files as well
(Makefile and reinst.sh).