lines2insns: Handle the kernel image too
Before this commit, lines2insns could handle only modules but not
vmlinux files. This did not allow to deal with the built-in modules
either.

Now the .text section of the kernel (vmlinux) is handled too. .init.text,
etc., are not feasible to process at the moment. In addition, the order
of the sections in memory is different from that for the modules.

One should consider removing the common headers from the input file if
the goal is to analyze a given built-in module only. For such headers,
lines2insns will find the insns in the kernel as a whole, which may be
too much.