lines2insns: Added a workaround for the weird behaviour of libdwfl
When used for some kernel modules, libdw/libdwfl 0.159 behaves in
a way different from what one would expect. As a result, one
cannot use dwfl_module_relocate_address() to find the section the
address belongs to, like eu-addr2line does, for example. The
function may not find the section because the addresses of
__ksymtab_gpl, __kcrctab, __kcrctab_gpl sections in the array of
sections (internal to libdwfl) are not relocated properly.
Or, perhaps, something else is needed for libdw to properly load
the module and avoid this problem - not sure what exactly.

So we just search the our array of sections explicitly.