Commit the state of the project before rewriting with Kprobes
This commit is to specify the state of RaceHound (not yet working though)
before I reimplement handling of software breakpoints using Kprobes.

This state of the code is saved in case it is needed in the future for
some reason.

By the way, here is an example of using Kprobes to execute a pre- and a
post-handler for an insn in the same context as the insn w.r.t. the
local interrrupts and preemption:

https://abf.io/spectre/kernel-examples/tree/master?path=kprobe_lite

Using Kprobes instead of handling the software breakpoints directly
should simplify the code (some its parts have become very similar to
Kprobes recently). It should also help reduce the number of hacks needed
to get all that working.

The technique in the example mentioned above does not use single-step
and trap flags, so, theoretically, it should co-exist with the hardware
breakpoints nicely. We'll see.