Overhauled the architecture of RaceHound, rewrote its key parts
RaceHound can now be used to watch any instruction where a Kprobe can
be placed (if the instruction accesses non-stack memory, of course).
No matter if it is the code of a loadable module, a built-in module or
the kernel proper.

RaceHound is no longer restricted to watching the code of a single module
only.

It is now possible to start watching the instructions of the modules
that were already loaded at the time RaceHound was loaded.

If the target kernel modules are unloaded and then loaded again,
RaceHound will now revalidate the software breakpoints it set for these
modules before and will automatically set the ones that are valid. No
need to reload RaceHound or reset the breakpoints manually in that case.

To add a software breakpoint, an appropriate string should be written to
'racehound/breakpoints' file in debugfs.

The format is "[<module>:]{init|core}+0xoffset", the same as lines2insns
tool uses.

Examples:
  test_module_simple:core+0x25
  sample_target:init+0x153

If module name is not specified, the breakpoint is for the kernel proper
or a builtin module. Example:
  core+0x2a4e23

To remove a breakpoint, write the same string but with '-' prepended to
that file.

"init" and "core" areas.
Both contain ELF sections of the target with flags
SHF_ALLOC | SHF_EXECINSTR set, in the same order as in the target,
without gaps inbetween. "init" area contains such of these sections that
have names starting with ".init", "core" area contains the rest.

This is similar to what the module loader uses in the kernel (see
module_init and module_core in struct module).

Kprobes are now used to take care of the software breakpoints. Among
other benefits, this allowed to remove a whole lot of not-so-obvious
code from RaceHound. Some tricks are used to make sure the appropriate
handler runs before the instruction of interest in the same context as
the latter. See the previous commit for details.

It is now also possible to set different delays for atomic and
non-atomic context, see the module parameters, "delay" and
"delay_in_atomic". If the latter is not set, "delay" will be used for
both.

Randomization of the breakpoints as well as everything related to the
"expiration" of breakpoints was removed from the kernel part of
RaceHound. Such things may be implemented in the user space in the
future. I suppose, kernel is not the proper place for this.

The kernel part of RaceHound (racehound.ko module) now only sets and
removes the breakpoints it has been told to, looks for the races and
reports them.

The tests and Readme are yet to be updated.