core: Do not monitor accesses to the stack
I guess, races on the stack variables should be rare. The accesses to
the local variables, on the other hand, are not always filtered out by
ma_lines+lines2insns.

Let us detect if the insn is about to access something on the stack and
ignore such accesses.

While at it, I made some cosmetic changes as well.