lines2insns: Implemented section-to-area and the reverse conversions
Convenience feature. May help when the locations of the insns to be
monitored come, say, from KEDR reports as module:section+0xoffset.

lines2insns --section-to-area will convert that to
module:{init|core}+0xoffset_in_init_or_core, which is what the kernel
part of RaceHound expects.

The appropriate binary file (the module or the kernel image) is needed
for such conversions.

lines2insns --area-to-section performs the reverse conversion.