lines2insns: Use the correct path to the source file for comparisons
Should fix https://github.com/winnukem/racehound/issues/7.

It is more reliable to use the source path stored in the DWARF info
rather than the one passed by the user. They may be different. For
example:
  /home/tester/work/kernel/drivers/media/usb/uvc/uvc_video.c
is in the DWARF info, and the user may pass just
  uvc_video.c

As a result, lines2insns failed to understand this is the same source
file. This commit fixes the issue.