core: If the kernel defines KPROBE_INSN_SLOT_SIZE, use it
The actual size of the insn slots may be greater than MAX_INSN_SIZE in
some kernels. If so and if the kernel defines KPROBE_INSN_SLOT_SIZE, we
can use it.

This may allow to monitor the insns of 11 bytes in size, e.g., some
forms of "mov 0x0, <somewhere>", which are quite common.