Allow different delays for different instructions
It is often useful to make different delays for different monitored
instructions. Before this commit, that required to reload racehound.ko
each time with different "delay" parameter. Now it is possible to
specify a delay for a BP when adding this BP via racehound/breakpoints
file.

The format is now as follows:
  [<module_name>:]{init|core}+0xoffset[,delay=<value>]

The value of "delay" is in milliseconds. Unless it is 0, it is used
instead of "delay" and "delay_in_atomic" parameters of racehound.ko.