Added racehound/events file that lists the hit BPs
The user-space apps can poll it to be notifies when the new events are
available there, that is, when the new breakpoint hits happen.

The list of the hit breakpoints can then be read, one per line. The
format if the same as it is for racehound/breakpoints.

All this can be used to prepare user-space applications that use
RaceHound to place or remove breakpoints according to some policy. See
DataCollider for MS Windows for the examples of such policies.