examples: Added grouping of race reports in manage_bps.py
Added -q/--quiet option as well. It suppresses the normal output of
events. Only the summary of the found races will be output when the
script exits.

With all this in place, the script can be used to monitor how RaceHound
works and see which races it finds without digging through the system
log.