Added check_races.py, a tool to sweep through the code
check_races.py gets the list of the instructions to monitor and
instructs the kernel part of RaceHound to add/remove breakpoints as
needed for this.

The policy is similar to what MS DataCollider does: if the BP hit rate
is too low, we add more BPs, if it is too high, we don't. If a BP has
been hit, it is removed and another one (if available) is placed,
randomly chosen from the list of the instructions to be monitored.