tests: Generate the lists of breakpoints to set using lines2insns
...rather than hard-coding them. Different distros / compilers / other
things result in the instructions of interest having different addresses
in the test kernel modules. Keeping the lists of breakpoints up-to-date
manually is tedious and time-consuming.

However, it is the particular source lines we are interested in during
the tests. Let us keep the lists of these source lines instead and
generate the lists of breakpoints from these using lines2insns.

I also removed the separate test for lines2insns, it became meaningless
after this change (it checked that the breakpoints lines2insns found
were present in the lists). Instead, the tool will be checked:
* partially during the build process of the tests - the process will fail if
  lines2insns finds nothing,
* partially during other tests that use the generated lists of breakpoints -
  if the generated lists are wrong, RaceHound may not find (some of) the
  races and the tests will fail.