Revisited reporting of the races: output the call trace, etc.
The locations of the machine instructions that raced on a memory block
are often not enough to diagnose the problem. I added output of call
traces for the conflicting threads, where available.

To avoid spamming the system log with tons of backtraces of identical
races, the race reports with the same pair of call traces are output
only once, that is, for the first occurrence if the race only.

stackdepot is used to store the call traces (I copied it from the
mainline kernel because not all the target kernels have it atm) and the
information about the found races is kept in a hash table.

Reporting the races may take some time, so I moved it to the process
context: to the workqueue items. Needlessly spending time in the
handlers of software and hardware breakpoints is usually a bad idea. A
downside of the deferred reporting is that some stack trace entries can
be resolved incorrectly. For example, this may happen if an entry belongs
to an init area of a kernel module or to the module that has been unloaded
before reporting starts. For now, this should not be much of a problem.

While at it, I removed reporting the race events to user space via the
events file. If the log is not spammed with race reports, it may be
easier to use just that like other debugging tools do.

Note that the identical races may still be reported several times, e.g. if
found with HW BP and later - with repeated read. I have also observed
the kernel not showing some of the stack trace entries, which also makes
the traces different and causes RaceHound to think the race is new.

Not sure why kernel skipped these trace entries, will take a look later.

Anyway, these are still some issues with this new reporting scheme. They
should be addressed later.